+++ src/models/analytics-query.ts
@@ -55,14 +55,14 @@ const conditionSchema = Joi.object().keys({
 const schema = Joi.object().keys({
   source: Joi.string().example('pinGroup').required(),
   columns: Joi.array().items(Joi.alternatives().try(
-    Joi.object().keys({ field: fieldSchema, name: Joi.string(), dataTypeHint: Joi.string().valid('string', 'integer', 'number', 'boolean') }).required(),
+    Joi.object().keys({ field: fieldSchema, name: Joi.string(), dataTypeHint: Joi.string().valid('string', 'integer', 'number', 'boolean', 'hashId') }).required(),
     Joi.object().keys({ type: Joi.string().valid('count', 'share').required(), condition: conditionSchema, name: Joi.string() }).required(),
     Joi.object().keys({
       type: Joi.string().valid('sum', 'mean', 'min', 'max', 'any').required(),
       condition: conditionSchema,
       field: fieldSchema,
       name: Joi.string(),
-      dataTypeHint: Joi.string().valid('string', 'integer', 'number', 'boolean'),
+      dataTypeHint: Joi.string().valid('string', 'integer', 'number', 'boolean', 'hashId'),
     }).required(),
     Joi.object().keys({
       type: Joi.string().valid('timeGroup').required(),
@@ -128,12 +128,12 @@ type AggregatedColumn =
     condition?: Condition;
     field: Field;
     name?: string;
-    dataTypeHint?: undefined | 'string' | 'integer' | 'boolean' | 'number';
+    dataTypeHint?: undefined | 'string' | 'integer' | 'boolean' | 'number' | 'hashId';
   };
 
 type TimeGroupColumn = { type: 'timeGroup', field: string; granularity: TimeGranularity, name?: string };
 
-type UnaggregatedColumn = { type: undefined, field: Field, name?: string, dataTypeHint?: undefined | 'string' | 'integer' | 'boolean' | 'number' };
+type UnaggregatedColumn = { type: undefined, field: Field, name?: string, dataTypeHint?: undefined | 'string' | 'integer' | 'boolean' | 'number' | 'hashId' };
 
 type Column = UnaggregatedColumn | AggregatedColumn | TimeGroupColumn